Fixed tag list serialization when null.

+++ src/modules/rest/extjs/src/main/java/it/geosolutions/geostore/services/rest/impl/RESTExtJsServiceImpl.java
@@ -724,7 +724,7 @@ public class RESTExtJsServiceImpl extends RESTServiceImpl implements RESTExtJsSe
         if (tags == null) {
             return new TagList();
         }
-        return new TagList(tags, tags.size());
+        return new TagList(tags, (long) tags.size());
     }
 
     /**
